requirements.txt
appdirs==1.4.4
asgiref==3.4.1
black==21.7b0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
Deprecated==1.2.12
Django==3.2.5
django-oauth-toolkit==1.5.0
django-service-objects==0.7.0
djangorestframework==3.12.4
idna==3.2
isort==5.9.2
jwcrypto==0.9.1
mypy-extensions==0.4.3
mysql==0.0.3
mysqlclient==2.0.3
oauth2-provider==0.0
oauthlib==3.1.1
pathspec==0.9.0
pycparser==2.20
pytz==2021.1
regex==2021.7.6
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
tomli==1.1.0
urllib3==1.26.6
wrapt==1.12.1
